+++ docs/main.js
@@ -5,7 +5,6 @@
  * @property {module:tasks/checkFileNames.Parameters} [checkFileNames] Parameters of the checkFileNames task
  * @property {module:tasks/jsdoc.Parameters} [jsdoc] Parameters of the jsdoc task
  * @property {module:tasks/nsp.Parameters} [nsp] Parameters of the nsp task
- * @property {module:tasks/jscs.Parameters} [jscs] Parameters of the jscs task
  * @property {module:tasks/eslint.Parameters} [eslint] Parameters of the eslint task
  * @property {module:tasks/jest.Parameters} [jest] Parameters of the jest task
  * @property {module:tasks/mocha.Parameters} [mocha] Parameters of the mocha task
@@ -23,7 +22,7 @@
  * @property {Object} typescript Common typescript compiler configuration, used by:
  *                          babel, jest, jsdoc, check-dependencies, mocha
  * @property {string|string[]} sourceFiles Common source files, used by:
- *                                         eslint, jscs, check-dependencies, check-file-names
+ *                                         eslint, check-dependencies, check-file-names
  * @property {boolean} allowLinking Whether to allow linking globally installed packages, used by:
  *                                  npm-install
  */