+++ src/tasks/jscs/__tests__/default-configuration/src/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-export function sum(...args:number[]):number {
-  return args.reduce((acc, val) => acc + val, 0);
-}
-
-export async function sumAsync(...args:number[]) {
-  return sum(...args);
-}