+++ src/tasks/jscs/__tests__/fix-task/gulpfile.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const gulp = require('gulp');
-const task = require('../../index');
-
-task.dependencies = {};
-task.register(gulp, 'jscs', {
-  files: 'index-invalid.js'
-});